[tests] Added always_scale argument (boolean) to tests_default_random() so that when it is true, scaling is always performed on random numbers generated by mpfr_urandomb(), instead of preferring numbers of the order of magnitude of 1 for half outputs. This will be useful for mpfr_rint_* tests.
For tgeneric.c, the value of this argument is controlled by a new macro
TEST_RANDOM_ALWAYS_SCALE (0 by default, which corresponds to the current
behavior). For the other tests that call tests_default_random(), the
argument is set to 0, so that the behavior remains the same.

(merged changeset r9471 from the trunk)

[From SVN r9491 (branches/3.1)]