[README] Changed link to the README.dev file to the 4.1 branch.
It is probably better to give the link in the 4.1 branch, since
we are in the 4.1 branch: some changes in master may not apply
to the 4.1 branch (for instance, the list of macros and the list
of environment variables that affect the tests). Future changes
useful for the 4.1 branch could be cherry-picked if need be.