[tests/mpfr-test.h] Improved the change done a few hours earlier.
The code introduced by commit 558a246e612fa230623fb0dfa0570b1c626a103c
was changing the conditions under which IGNORE_CPP_COMPAT is defined
(to ignore -Wc++-compat with a GCC pragma). Let's simplify it by using
our __MPFR_GNUC() macro. The only possible change of behavior is that
this excludes Intel's compiler (which is probably better as testing
the GCC version on Intel's compiler doesn't make sense).

Note concerning the behavior of Intel icx 2021.3.0 under Linux: Before
the above commit, it was not excluded; but after this above commit, it
was excluded as it identifies itself as GCC 4.2.1 (like Clang). With
this new change, it is still excluded. Anyway, it ignores -Wc++-compat
and doesn't complain on the GCC pragmas, so the above choice currently
doesn't matter.

(cherry picked from commit ff4511247148eeb1255ffb8e9d4831d783610cd5)