[tests/tsprintf.c] Modified a buggy test of the thousands separator.
The test

  check_vsprintf ("+01,234,567  :", "%0+ -'13.10Pd:", (mpfr_prec_t) 1234567);

is based on the output from glibc up to 2.36, which is incorrect:

  https://sourceware.org/bugzilla/show_bug.cgi?id=23432

The GNU C Library has apparently been partially fixed in its Git
repository for the future 2.37, since a tsprintf failure has been
reported (this is a bug in this test, not in the MPFR library):

  https://sympa.inria.fr/sympa/arc/mpfr/2023-01/msg00001.html

So, modified the test to avoid the particular case of leading zeros
due to the precision field larger than the number of digits. This
case has already been tested without the thousands separator (where
there are no issues with the C libraries), so that we do not miss
much testing. Added a comment explaining the issue and a possible
solution for future testing of this particular case (if need be).

(cherry picked from commit 5172494c09718ffcb7ef1f19b3b211e3bce8781a)