[tests/tsprintf.c] Added a comment about a failure with glibc 2.37.
Klaus Dittrich reported in the MPFR mailing-list that the modified
test from commit 4f03d40b5a5ceeb3126dc51c3511bdceef3bab19 was still
failing with the glibc git. But this time, this is entirely a bug
in glibc, present in glibc 2.37:

  https://sourceware.org/bugzilla/show_bug.cgi?id=30068

with even a possible buffer overflow in sprintf (as I guessed
in Comment 2 and was confirmed later). → CVE-2023-25139

Thus the test still fails with glibc 2.37, and the future fix of
the glibc vulnerability should also fix this failure.