Fixed possible stack overflow in mpfr_rec_sqrt + non-regression test.
* src/rec_sqrt.c: use a standard Ziv loop.
* tests/trec_sqrt.c: added a non-regression test with bad_cases().

Details:

The issue is due to many allocations in the stack (with alloca(),
via MPFR_TMP_LIMBS_ALLOC) on cases that are very hard to round (from
an analysis on a testcase given by Fredrik Johansson, not included
here) with a not too large destination precision: many iterations
are needed due to a specific Ziv loop where the working precision
is increased additively, and it is not possible to free the data
allocated in the previous iterations.

The fix is to use a standard Ziv loop, as it increases the working
precision multiplicatively instead of additively, so that it needs
much fewer iterations in such cases. This should also make the code
much faster in such cases.

This issue was not detected by bad_cases() in tests/trec_sqrt.c
because to be able to reproduce it, one needs a much larger precision
than the ones tested with the chosen parameters.

A new bad_cases() instance, with a large enough precision for the
input (py = 9999, psup = 120000, giving px = py + psup = 129999),
is added here in order to be able to reproduce the failure.

The testcase given by Fredrik Johansson is not included here because
it uses a very large string for the input (the point of bad_cases()
is to generate such an input automatically).

Cherry picked from commits
  3846a4105bc5e5b9965fb8dc618d668817a8be08
  270f4df6b3a49caae1cf564dcdc1c55b1c5989eb