[src/reldiff.c] Fixed mpfr_reldiff on special values.
The mpfr_reldiff function, which computes |b-c|/b, was buggy on
special values, e.g. on the following (b,c) values:
  * (+Inf,+Inf) gave ±0 instead of NaN (like NaN/Inf).
  * (+0,+0) gave 1 instead of NaN (like 0/0).
  * (+0,1) gave 1 instead of Inf (like 1/0).

The sign of 0 for (+Inf,+Inf) or (-Inf,-Inf) was not set, i.e. it
was just the sign of the destination before the call. This made
the "reuse" test fail after its update to check the sign of zero
(commit e6d47b8f5ab666e658e05fe282e3490a06c0278e), as the results
were not consistent.

For b = 0, there was a comment "reldiff = abs(c)/c = sign(c)", which
mixed up b and c (one should divide by b, not by c), and this could
explain some bugs.

Note that this comment (and associated code), if corrected, seemed to
imply that for c = 0 and b finite (and non-zero?), the intent was to
return sign(b). However, this does not follow the documentation, which
says that the formula is just computed using the precision of the
destination and the provided rounding mode for all operations. So the
final result may be affected by intermediate roundings, while sign(b)
assumes that the function is correctly rounded, which is not the case
for this function (just there for compatibility with MPF). Deviating
from the documentation in some cases could break some properties (like
monotonicity?).

These bugs were not found because mpfr_reldiff is not tested at all,
except by "reuse", which just checks some consistency when a variable
is reused for the destination.

Cherry picked from commits
  9862e37b18d68187c309d73a2d6f86f3ab3186a2
  ac3007bbad422358a3b3b3ae6448fe2f795641ed