[tests.c] In bad_cases(), handle the inex_inv case better.
The "f exact while f^(-1) inexact" case is actually almost the same
as the "inverse doesn't match" case. So, handle it in the same way,
avoiding a fatal error (which would be bad for the non-developers):
a few occurrences of this case should be regarded as normal, and if
this occurs too often, this will be detected like the other case
when the MPFR_CHECK_BADCASES environment variable is set.

This is useful for future bad_cases() calls.

Cherry picked from commits
  be6c78fa0adde7fdcdac7e9ada672b90ed8fcbf5
  0d64e25ceb23089c818f3a9a3976a088126705cf