[tests/inp_str.dat] Make the word that contains '\0' look like a number.
Changed ".\0." to "1\01" (\0 being the null character) in order to
trigger a failure with the current mpfr_inp_str code. The issue is
that '\0' makes the word look like the string "1", the '\0' being
regarded as a terminating null character. This is incorrect as '\0'
is part of the full word.

(cherry picked from commit c9523bc6fb11c8eefc75524e947e5eb81e682675)