[src/mpfr-gmp.h] Fix integer overflow when mp_limb_t < int (mini-gmp).
When mp_limb_t is shorter than int (which is possible with mini-gmp),
an integer promotion in udiv_qr_3by2 triggers an integer overflow,
detected by testing MPFR with

  ./configure --with-mini-gmp=/path/to/mini-gmp CC=gcc-13 \
  CFLAGS="-O2 -fsanitize=undefined -fno-sanitize-recover
          -DMINI_GMP_LIMB_TYPE=short"

(many tests fail).

Note that this is not detected by GCC 12 and before, probably due to
an optimization bug that hides the UBsan error.

Moreover, mini-gmp currently has the same bug in gmp_udiv_qr_3by2,
but also in gmp_udiv_qrnnd_preinv; it needs to be fixed too in order
to avoid the failures in the MPFR tests.