[tests/trint.c] C2x compat: replaced function without a prototype.
The second parameter of test_fct was "int (*g)()", i.e. a pointer to
a function without a prototype, because 2 kinds of prototypes can be
used:
  * one with args (mpfr_ptr, mpfr_srcptr);
  * one with args (mpfr_ptr, mpfr_srcptr, mpfr_rnd_t),
    but only for mpfr_rint.
This was not a problem thanks to the test "g == &mpfr_rint", i.e.
we knew which one to use. However, as the Clang 16 warning message
says for -Wdeprecated-non-prototype:
"passing arguments to a function without a prototype is deprecated
in all versions of C and is not supported in C2x".

The best way to fix this is to use the prototype with 2 arguments
(mpfr_ptr, mpfr_srcptr) for the parameter, as this is the general
case. All we need to do is to cast &mpfr_rint to this prototype
(this appears only twice).

As these casts trigger the -Wcast-function-type warning when enabled,
let's explicitly disable this warning for GCC and Clang in versions
where it has been introduced.

Cherry picked from commits
  019e961a9598ab0a475cc4e5dd7eba17f3546f51
  235029b5ccbc54910154892b7c2c566e85043bae
  4a7105e807f02bc102f31c42bb922d00f1a1fc14