[acinclude.m4] Replaced <gmp.h> by "gmp.h" for #include.
The tests from configure.ac all use #include "gmp.h", and for the
MPFR build and tests, gmp.h is also included with #include "gmp.h"
(by mpfr-impl.h). For consistency, do the same in acinclude.m4.

Note: In practice, there should not be a difference, except if there
is a gmp.h file somewhere in the MPFR build tree. But this is not
supported, and this may break the build or the testsuite (whatever
is used in the configure tests).

(cherry picked from commit 1908c65a4ac9f62cb560edf7324ace33a66661d8)