[configure.ac] For libgmp, replaced AC_CHECK_LIB by our own link test.
The issue with AC_CHECK_LIB is that it cannot know the prototype of
mpz_init. Thus it declares the function without a prototype, which
is obsolescent in ISO C, so that the compiler may complain (e.g.
with GCC's -Werror=strict-prototypes option, which may be useful to
detect bugs in MPFR). Moreover, the call is done with an incorrect
prototype, which could also be an issue on some systems (currently,
no prototype check is done when linking, but this might change in
the future).

Note also that this new test is more robust, in case GMP changes the
symbol name in the future (the API does not need to be changed).

For the reference:
  https://sympa.inria.fr/sympa/arc/mpfr/2023-08/msg00003.html

(cherry picked from commit e986f21c256c38b31f9cf3e564f779c4405d0767)