Formatted output functions: fixed ‘+’/space flags on NaN and Inf.
This is about the documentation and behavior of the mpfr_printf,
etc. functions on MPFR numbers (mpfr_t) only, i.e. with the ‘R’
type specifier.

The ‘+’ and space flags were ignored on NaN and Inf. While this
was loosely documented as such (without an explicit mention of
these flags), the GNU MPFR manual also says that the flags have
the same meaning as for the standard ‘printf’ function. So this
was contradictory and regarded as a bug.

Behaving like the ISO C standard should give less surprise, and
this is probably what is expected (better for alignment purpose).

This was partly discussed at
  https://sympa.inria.fr/sympa/arc/mpfr/2023-03/msg00003.html
(only for NaN and the ‘+’ flag at that time).

* doc/mpfr.texi: corrected the documentation.
* src/vasprintf.c: fixed (the code is now simpler since the rule
  is the same in all cases: real and non-real values).
* tests/tsprintf.c: grouped all the tests of NaN and Inf, and
  added much more tests (by constructing the format string in
  nested loops instead of just a few hardcoded cases).

Cherry picked from commits
  0ab45735b1b6d29a7e824ecdef4c27e71b2eca65
  6c221eed3d0ffd2fb904b9a9e0e62e0fa4ba57da