[strtofr.c] Fixed 2 mpfr_strtofr bugs in case of huge negative exponent.
This fixes the bugs triggered by bug20230606, mentioned at
  https://sympa.inria.fr/sympa/arc/mpfr/2023-06/msg00000.html
  https://sympa.inria.fr/sympa/arc/mpfr/2023-06/msg00001.html

These bugs are integer overflows, meaning undefined behavior.
In practice, the consequences may be incorrect results. But
for the first bug, it has been seen that a GCC optimization
makes the bug invisible.

Note: There are other issues with the code for huge exponents
(in strtofr.c and the MPFR_SADD_OVERFLOW macro from mpfr-impl.h).
See FIXME from commits
  aa30bfe3c868640852fa2510b0f228170e39c0c5
  e06a073c157fe95dd24d16d9fc3eef3781b58795
in master. But it is not clear whether the problematic cases can
occur in the context of mpfr_strtofr: my attempts to trigger a
problematic case failed, so that there is a chance that the code
is actually correct; but many parameters are involved, including
those based on the internal structure in limbs, so that it is
difficult to say for now. A closer analysis needs to be done.
Even if correct, the code would need to be cleaned up, proved,
and better documented.

(cherry picked from commit 007c66d8dbe9ac5acbd7e245bb01af5df968044b)